feat(debug-ui): extend epoch info api (#12891)
Allows to query arbitrary epoch by id in the epoch validators page.
The interface allows to either input id or scroll back and forth between
epochs.

The main caveat is:
* current epoch must be indexed by `ValidatorInfoIdentifier::BlockHash`
because it wasn't finalised yet and it is a sign that its stats are
being updated online;
* for past epochs we must use `ValidatorInfoIdentifier::EpochId` because
`BlockInfo` data used for block navigation may be GC-d.

Other small fixes:
* Support NotEnoughChunkEndorsements kickout reason
* Display kickout among the roles for the next epoch because that's
where it actually comes in place. Displaying it for the current epoch is
confusing.

See video preview here
https://near.zulipchat.com/#narrow/channel/295558-core/topic/Optimistic.20Block.20Design/near/498834846