forknet: add test for forknet v2 setup (#12950)
It's kind of hard to test changes to the fork-network commands at the
moment, so this adds at least a quick check that the forked network can
start and produce some blocks. It would be good to add a check that we
can send mirrored transactions, but that is prob not a huge deal to
cover for now.

This also makes a change to `Key.from_keypair()`, because creating a key
with `Key.from_random()` and then saving it to a file with
`Key.to_json()` does not work because it only writes the bytes of the
public key, but neard expects the key to be the private and public keys
concatenated. This should be fine to change because `Key.sign_bytes()`
does make sure to take the first 32 bytes