fix(resharding): Refcounts quick fix (#12998)
A quick mitigation for refcount issues after resharding.

In flat storage postprocessing, we duplicate values (leaves) that are
copied to both children.
However, that alone does not make the repro
([link](https://github.com/near/nearcore/pull/12993)) pass because the
repro does not use memtries, and `TrieKey::PromiseYieldIndices` (`10`)
has an intermediary node.

The intermediary node is accessed during the memtrie split and is
already recorded.
We can bump refcounts for all recorded nodes (whose number is bounded by
the trie depth), potentially leaking them in the worst case. However, it
is possible that we should duplicate all of them anyway.