fix(optimistic_block): set catchup mode correctly (#13014)
If node is catching up some shard, we need to pass `NotCaughtUp` even
though we are not going to do "optimistic catchup".

The reason is that `IsCaughtUp` assumes that all shards we care about
are already caught up, tries to read chunk extra which doesn't exist yet
and fail to process optimistic block.

The fix is unfortunately hacky but it doesn't impact much and I'm in a
bit of a rush before 2.6.