[testloop] Introduce identifier in testloop callback event (#13016)
Within TestLoop framework, we keep a priority queue of events to be
executed. As of today there's no way to distinguish events of one client
from another.

This PR introduces a new `identifier` field within the event. This is
set to be the string representation of `account_id` of the client.

In the future this would be useful to implement node shutdown and
restart functionality within TestLoop where we would potentially be able
to filter events based on the identifier.

Couple of other cosmetic changes to test loop.