benchmark_mpc_sign: write nonces even if some txs error (#13022)
Before this change, the synthetic benchmark script exited immediately
upon encountering an RPC error. It meant that the nonces tracked in
memory did not get written to disk, leading to nonce conflicts when
running the script again.

This PR fixes that issue by gracefully handling errors and always
updating the nonces on disk.