[testloop] Support to kill a node (#13038)
This is the first step for adding support to restart nodes in testloop.

Here we begin by providing the support to kill a node.

TestLoop visualizer is updated appropriately to show events to a killed
node in red.


![image](https://github.com/user-attachments/assets/205349fa-6302-4e3b-8960-230bf9822358)

Next step would be to add support to start/restart a node