Introduce `ValidatedTransaction` (#13046)
This PR introduces the `ValidatedTransaction` type and starts to use it
in `process_tx_internal` i.e. when a transaction is received from the
rpc nodes. In order to keep the scope small enough, this PR does not
store validated transactions in the tx pool yet so we still have to do
validation when we produce chunks.

---------

Co-authored-by: Simonas Kazlauskas <github@kazlauskas.me>