test: add validator_rotation example for testloop. (#13147)
Some tests depend on validator assignment rotation between epochs and
currently are using mock epoch manager for that. In order to migrate
them off of mock epoch manager (as part of
https://github.com/near/nearcore/issues/10634) we would need to change
validator assignment in-between epochs. The best candidate for those
tests is testloop so this adds example on how to rotate validator
assignments in testloop.

Sample tests depending on validator assignments between epochs:
*
https://github.com/attila0x2A/nearcore/blob/b47b467a49991bca20416e52b31d8b738f0f03b6/integration-tests/src/tests/client/consensus.rs#L36
* Most tests depending on
[`setup_mock_all_validators`](https://github.com/attila0x2A/nearcore/blob/b47b467a49991bca20416e52b31d8b738f0f03b6/integration-tests/src/env/setup.rs#L825)