feat(database): drop column cmd (#13061)
Running a newer version of the binary may cause a database migration by
adding a new column.
If the new binary is run by mistake, the desired (older) binary will no
longer work because it does not recognize the newly added column.

We encountered this issue
([link](https://near.zulipchat.com/#narrow/channel/308695-nearone.2Fprivate.topic/2.2E5.2E0.20-.20TestNodes.20got.20stuck/near/503050524)),
but fortunately, it did not affect the recent mainnet release—simply
because this particular release did not perform any database migrations.

Having a neard command to remove a column from the database would be
helpful.

Note – In this PR, new columns can only be removed by the same binary
that performed the migration, as the old binary cannot resolve the
column names. Additionally, the database version needs to be restored,
which is implemented in a follow-up PR:
[link](https://github.com/near/nearcore/pull/13082).