feat(database): reset database version (#13082)
Follow-up to https://github.com/near/nearcore/pull/13061

When a migration adds an extra column, simply removing it is not enough
to revert the migration. We also need to restore the database version to
the previous one. This PR introduces a command that allows us to do that
using the target binary.