[genesis] Separate mainnet/testnet genesis initialization (#13158)
With limited replayability in sight, we would like to try to remove
older protocol versions from neard binary over time. One of the key
places where an old protocol version is used is the genesis protocol
version for mainnet and testnet, which is set and hardcoded as 29.

Note that forknet runs override and change the genesis protocol version
to be the latest. We don't expect any other place to use genesis
protocol version 29 other than mainnet and testnet.

We read and initialize the genesis during neard startup. Currently the
code path for genesis block and genesis chunk is shared with the code
path for latest block and chunk production.

This PR does a couple of things to separate out handling genesis
protocol version 29
- Add hardcoded checks for genesis hash to be either from mainnet or
testnet
- Add hardcoded checks for state roots for mainnet and testnet.
- Remove genesis json hash altogether as it doesn't have much utility
any more when we are able to check state roots.
- Separate out block and chunk code for protocol version 29
(PROD_GENESIS_PROTOCOL_VERSION) so that we don't have to go through the
`protocol_version.enabled()` codepath.