fix: chain fixes for tx double processing (#13164)
1. Before scheduling apply chunks for optimistic block, check that block
is not in processing already. We thought that tip check is enough, but
for TPS-intensive tests it's possible that block and optimistic block
may be scheduled close in time. Call to `is_height_processed` does
exactly that. Extending `test_process_optimistic_block` to cover it and
checking that it fails without this call.

2. Also, sending endorsement to self bypassing the network, if we are
the block producer. There was a side effect of
`test_client_with_simple_test_loop` never stopping processing events due
to that. I think it is reasonable to stop processing events explicitly
as soon as test loop test was shut down. It seems that before we were
relying on the fact that at least one network communication is required
to produce new block; now we are removing it.