refactor: remove a clone for `SignedTransactions` (#13261)
Part of https://github.com/near/nearcore/issues/13140. 

Similar to https://github.com/near/nearcore/pull/13122 which can be
closed in favour of landing this.

- `SignedValidPeriodTransactions` take ownership of `SignedTransactions`
- Rework `ApplyProcessingReceiptState` and `ApplyProcessingState` to not
store `SignedValidPeriodTransactions`
- This allows us from avoiding a clone in `process_transactions`. We
need to do a clone in `set_state_finalize` instead now. However that is
not part of chunk application but instead part of state sync. And still,
in this case we are not introducing a new clone just moving it to the
top of the stack.

CC: @miloserdow. As discussed in our chat, this should help with your PR
and hopefully we will not have any clones in your work.

---------

Co-authored-by: Simonas Kazlauskas <git@kazlauskas.me>