test: rewrite consesus integration test in test loop. (#13244)
Part of https://github.com/near/nearcore/issues/10634

There is only one integration test for consensus that is depending on
mock epoch manager for validator rotation:
`ultra_slow_test_consensus_with_epoch_switches`. In this PR I rewrite it
for testloop, but I am trying to keep most of the implementation the
same to make it easier to compare with previous version.

Since this test required validator rotation I have wrote a utility to
make it easier to run test loop with rotating validators and rewrote
related example to use it.

The test itself is probabilistic and tries to delay and skip some of the
blocks to test for finality bugs.

While deflaking the test I found one minor issue of denominator == 0 in
chain/epoch-manager/src/validator_selection.rs which should affect only
tests and I fix it in this PR as well.