increase the epoch-length in create-delete-account test (#13330)
The small epoch lengh leads to some flaky behavior with
`do_create_account` leading to unrelated failures. Those include
producing one block with missing chunks due to
`ChunkHeaderReadyForInclusion` never called for that particular block in
the https://github.com/near/nearcore/pull/13259 . It has been verified
that the stress tested network does not produce more blocks with missing
chunks if the above PR is included. Together with the failures being
only observable with only small epoch lengths (5,6) and for only a
single block we have a strong evidence the failure is a noise.
Increasing the epoch length for the tests makes the test stable against
the changes.