feat: remove pessimistic gas pricing (#13370)
With `ProtocolFeature::ReducedGasRefunds` enabled, we remove the
pessimistic gas price increase. This PR makes this change for nightly
only.

This is the first part for NEP-536, which aims to reduce the number of
gas refund receipts that are generated in the system. Pessimistic gas
pricing is one major cause for such refunds, as it forces all non-local
receipts to pay more gas than it likely needs.

# Testing strategy

There is already a number of integration tests that had to be updated,
since the number of receipts has changed. (The refunding receipts is
missing with the feature enabled.) This should be enough to test the
basic functionality.

The other features for NEP-536, which adds a fee for all remaining gas
refunds, will require new integration tests which will also indirectly
test that pessimistic gas pricing no longer applies.