feat: do not refund gas price changes (#13418)
As part of NEP-536, pessimistic gas pricing is removed.
However, prices are still going up and down, causing differences in
purchase value and burn value.

It seems overall simpler for everybody, if we just ignore the gas delta
and write off the difference. In other words, charge the user the price
the paid for when purchasing the gas, instead of the latest block's
price.