fix flaky pytest upgradable.py (#13466)
This test was flaking in two different ways:

- `wait_epoch` waits for one particular node to reach the new epoch.
Then, we immediately call `assert_protocol_version` and check the
protocol version on all of the nodes. Another node may be in the process
of transitioning to the new epoch and need a little more time.

- `num_expected_endorsements` and `num_produced_endorsements` may not
update synchronously. Again, allowing a little time for them to match
resolves the issue.

In my local setup without the fixes, I observed the test flake in 2 of 3
runs. After these changes it ran successfully 5/5 times.