feat: pool of blocks pending optimistic execution (#13139)
Sometimes node can receive block during the processing of optimistic
block. Currently we ignore optimistic block at all, but it means that
chunk apply work for optimistic block is throwaway and there can be
performance losses.

Evidence that this case appears: [#core/performance-optimization >
traces analysis of overloaded network @
💬](https://near.zulipchat.com/#narrow/channel/497211-core.2Fperformance-optimization/topic/traces.20analysis.20of.20overloaded.20network/near/513876291).
Later in the thread there is evidence that it is helpful.

Now, if we receive block during optimistic block is processing, we put
it to `PendingBlocksPool`. In case of forks it can have many blocks.

Ideally we’d have some pipeline for block processing. For now, I solve
it following the current logic. Block processing returns
Error::OptimisticBlockInProcessing, which is later taken in a big match
statement and block is put to pending pool. After that, the event we are
waiting for, postprocess_optimistic_block, triggers block processing
again.

Testing:

* The behaviour doesn’t require components other than Chain, so I
implement two “test_pending_block”s which involve only chain.
* Unit tests for PendingBlockPool.