refactor: remove unnecessary `Arc` in various `get_chunk`  (#13497)
We have many different functions called `get_chunk`. Most of them are
looking up the chunk from the DB so do not need to unnecessarily wrap
the returned chunk inside an `Arc`. This PR tidies this up. The benefit
of not unwrapping a struct in Arc is that it makes it possible to
consume it without having to actually copy the data out of it.